fix compatibility for step time

+++ modal_training_gym/common/run.py
@@ -90,7 +90,7 @@ class TrainingRun(BaseModel):
     completed_at: int | None = None
     updated_at: int = 0
     duration_seconds: int | None = None
-    step_times: dict[str, dict[str, float | None]] | None = None
+    step_times: dict[str, dict[str, int | None]] | None = None
     substep_times: dict[str, dict[str, dict[str, float | None]]] | None = None
     # Terminal failure message (Ray driver error / exception) for a failed run,
     # so the cause is queryable from the record and shown on the dashboard even
